The topology can now shut itself down after certain number of packets emitted

+++ src/test/java/uk/ac/gla/atanaspam/network/NetworkConfiguratorBoltTest.java
@@ -126,7 +126,7 @@ public class NetworkConfiguratorBoltTest {
         // when the report is processed
         bolt.execute(reportingTuple);
         // then the report should be stored in the state under task 2
-        assertEquals((double) 1, bolt.state.getNumberOfPacketsDropped());
+        assertEquals((long) 1, bolt.state.getNumberOfPacketsDropped());
     }
 
     @Test
